backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0
groq==0.9.0
chromadb==0.5.0
sentence-transformers==3.0.1
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.3
networkx==3.3
gitpython==3.1.43
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.7.4
pydantic-settings==2.3.4
numpy==1.26.4
sse-starlette==2.1.0
